warning from converting DWORD bitsPerSample to native WORD storage
In the waveout structure specifications, bits per sample is stored as a 16-bit integer, so maybe the function should be reading from a WORD type rather than DWORD type.  Besides, it doens't look like the waveOut code is used much anymore, so in the worst case, this just fixes a warning.

+++ AziAudio/WaveOut.cpp
@@ -13,7 +13,7 @@
 #include <stdio.h>
 #include "WaveOut.h"	
 
-void WaveOut::BeginWaveOut(char *filename, WORD channels, DWORD bitsPerSample, DWORD sampleRate)
+void WaveOut::BeginWaveOut(char *filename, WORD channels, WORD bitsPerSample, DWORD sampleRate)
 {
 	if (waveoutput != NULL) return;
 	// Clear header
